Pass extra test for hashTable
Added extra test to handle collisions for removal. Specifically, if there was a collision on insert, one of the values would be stored in a location different from the index returned by the hashing function -- call that 'bumped value'. If we are to remove 'bumped value' later, it should not remove the value that is at the index returned by the hashing function.